+++ graph_cast/db/connection.py
@@ -48,6 +48,10 @@ class Connection(abc.ABC):
     def delete_collections(self, cnames=(), gnames=(), delete_all=False):
         pass
 
+    @abc.abstractmethod
+    def init_db(self, conf_obj: Configurator, clean_start):
+        pass
+
     # @abc.abstractmethod
     # def get_collections(self):
     #     pass
@@ -116,13 +120,7 @@ class WSGIConfig(ConnectionConfig):
 
 
 def init_db(db_client: ConnectionType, conf_obj: Configurator, clean_start):
-    if clean_start:
-        db_client.delete_collections([], [], delete_all=True)
-        #     delete_collections(sys_db, vcollections + ecollections, actual_graphs)
-        # elif clean_start == "edges":
-        #     delete_collections(sys_db, ecollections, [])
-    db_client.define_collections(conf_obj.graph_config, conf_obj.vertex_config)
-    db_client.define_indices(conf_obj.graph_config, conf_obj.vertex_config)
+    db_client.init_db(conf_obj, clean_start)
 
 
 def concluding_db_transform(db_client: ConnectionType, conf_obj):