+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.4] - 2023-08-30
+
+### Fixed
+
+- not more info level logging, only debug
+- 
+
+### Changed
+
+- in ingest_json_files(): ncores -> n_threads
+- in ingest_tables(): n_thread -> n_threads
+
+
+
 ## [0.11.3] - 2023-06-24
 
 ### Fixed